fix(core): widen raptorq source_blocks from u8 to u16
The ObjectTransmissionInformation constructor expects u16 for
source_blocks. Using u8 caused an unnecessary narrowing that could
silently truncate on future multi-block configurations. Widen to u16
in both production decode path and test mirror.

+++ crates/fsqlite-core/src/raptorq_codec.rs
@@ -384,7 +384,7 @@ impl SymbolCodec for AsupersyncCodec {
         // `SymbolCodec::decode` only carries `k_source` + `symbol_size`, not the
         // encoder's block partitioning. Reconstruct the same single-block object
         // geometry that the current encoder metadata can faithfully describe.
-        let source_blocks = 1_u8;
+        let source_blocks = 1_u16;
         let symbols_per_block = k_source.max(1);
         let object_size = u64::from(k_source)
             .checked_mul(u64::from(symbol_size))
@@ -399,7 +399,7 @@ impl SymbolCodec for AsupersyncCodec {
                 what: "symbol_size as u16".to_owned(),
                 value: symbol_size.to_string(),
             })?,
-            source_blocks,
+            u16::from(source_blocks),
             u16::try_from(symbols_per_block).map_err(|_| FrankenError::OutOfRange {
                 what: "symbols_per_block as u16".to_owned(),
                 value: symbols_per_block.to_string(),
